Avoid serial_write_blocking() warnings.

+++ src/hardware/agilent-dmm/sched.c
@@ -140,7 +140,7 @@ static int agdmm_send(const struct sr_dev_inst *sdi, const char *cmd)
 		strcat(buf, "\r\n");
 	else
 		strcat(buf, "\n\r\n");
-	if (serial_write_blocking(serial, buf, strlen(buf)) < strlen(buf)) {
+	if (serial_write_blocking(serial, buf, strlen(buf)) < (int)strlen(buf)) {
 		sr_err("Failed to send.");
 		return SR_ERR;
 	}
